refactor(extraction): split 1136-line extraction.rs into an extraction/ module
Pure code movement — no logic change. `extraction.rs` (1136 lines) becomes
`extraction/{mod,parse,prompt,graph,run,tests}.rs`, each a cohesive unit:
  - parse  — LLM-response cleaning + JSON parsing
  - prompt — input assembly, system prompt, few-shot, task registration
  - graph  — instance links, literal encoding, dedup, existing-instance reads
  - run    — retry harness + run_extraction driver
  - tests  — the full unit-test module (moved verbatim)

`mod.rs` re-exports each submodule (`pub use …::*`) so the public path
`perspectives::extraction::*` is unchanged — no edits needed in
extraction_e2e / extraction_test_support or other callers. Only visibility
change: `relation_predicates` is now `pub(crate)` (build_extraction_input in
prompt.rs calls it across the module boundary).

Verified: 26/26 extraction tests pass locally (19 unit + 7 e2e), no new warnings.
