+++ rust-executor/src/perspectives/extraction/parse.rs
@@ -1,0 +1,31 @@
+use super::ProposedInstance;
+
+/// Parse a raw LLM response into proposed instances.
+///
+/// Local models wrap JSON in reasoning/markdown noise, so we first strip the
+/// common wrappers (mirrors Flux `LLMutils.ts`): `<think>…</think>` blocks,
+/// ```-fences, and trailing commas. Then parse as a JSON array.
+pub fn parse_extraction_response(raw: &str) -> anyhow::Result<Vec<ProposedInstance>> {
+    let cleaned = clean_llm_json(raw);
+    let instances: Vec<ProposedInstance> = serde_json::from_str(&cleaned).map_err(|e| {
+        anyhow::anyhow!("extraction JSON parse failed: {e}; cleaned payload: {cleaned}")
+    })?;
+    Ok(instances)
+}
+
+/// Strip the reasoning/markdown noise local models add around JSON.
+fn clean_llm_json(raw: &str) -> String {
+    // 1. Drop <think>…</think> reasoning blocks (non-greedy, across newlines).
+    let think = regex::Regex::new(r"(?s)<think>.*?</think>").unwrap();
+    let s = think.replace_all(raw, "");
+
+    // 2. Drop code fences ```json / ``` (keep the inner content).
+    let fence = regex::Regex::new(r"```[a-zA-Z0-9]*").unwrap();
+    let s = fence.replace_all(&s, "");
+
+    // 3. Remove trailing commas before a closing } or ] (invalid JSON, common).
+    let trailing = regex::Regex::new(r",(\s*[}\]])").unwrap();
+    let s = trailing.replace_all(&s, "$1");
+
+    s.trim().to_string()
+}
